SDC-5999: Cluster pipeline start error if SDC_RESOURCES empty
This patch removes the need to have the tarred directory non-empty.

It also removes the test TestTarFileCreator.testCreateEtcTarGzFileDirIsNotReadable
because the method canRead() that the test depends on might have
different semantic on different platforms (as is document in Java doc).
It so happened that the test succeeded on my machine, but it wasn't able
to succeed on Jenkins due to the differences in platform.

Change-Id: Ib13536ebd55f156f9019bd2f7fb62845914ccf10
Reviewed-on: https://review.streamsets.net/10988
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>