SDC-7671. Support UPDATE operation in MongoDB Destination
**BREAKING CHANGE**

This adds REPLACE as an SDC OperationType for MongoDB.
Existing "UPSERT" OperationType is incorrectly treated as REPLACE in
MongoDB Destination.

"UPSERT" is no longer supported by OperationType in MongoDB Destination.
Instead, upsert is a stge config (flag) that can be used in conjunction
with UPDATE and REPLACE OperationType.

This means that existing pipelines will fail with Unsupported Operation
when an UPSERT code is sent to the stage. This is not something that can
be migrated with a Stage Upgrader since its not a stage config.

Existing pipelines should be migrated to use UPDATE or REPLACE and the
appropriate upsert flag.

**BREAKING CHANGE**

Change-Id: I2b6a9bbec4488c986625af48034f1deb39c8e84b
Reviewed-on: https://review.streamsets.net/11078
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>