SDC-7743: Stop propagating RPMS for legacy stage libraries
The way the "rpm" build works is that we run a special maven module
called rpm that scans the repository for all rpms and copies them to
special directories (one for RHEL 6 another for RHEL 7). This mentioned
code already have simple filtering to skip the special "rpm" profile
itself.

This patch extends the filtering ability to copy rpms only for those
stagelibs that are present in the main build. E.g. stages in the dist in
folder streamsests-libs. Any non-stage related rpms or legacy stage rpms
will be skipped.

Hence the user experience will be - if user needs the legacy stage lib,
they will have to go to our "legacy repo" and download it's tarball. And
then install it using the custom stage libs facility that we have there.
There won't be RPMs available for the legacy stages nor will the package
manager work with legacy repo. This user experience is consistent across
tarball, rpm and parcel installations.

We can build on top of it in future if we realize that more user
friendly approach is required.

Change-Id: I0b525e48bc9f01684783a56b3d26c90785dccb92
Reviewed-on: https://review.streamsets.net/11087
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Adam Kunicki <adam@streamsets.com>