SDC-7211: Directory origin - First File to Process should be available for all data formats
Talked to few folks offline, apparently there is a history with the file
formats there - originally we wanted to have the initial file to be
available for "all data formats but whole file" and hence we added the
triggeredByValue config option. Later on we've decided to enable those
config for whole file as well, but rather then dropping the whole
config, we just added WHOLE_FILE as an option. And then when I showed up
and updated the list of supported file formats, I haven't even realized
that there is a dependency. Hence this patch simply drops the whole
dependency declaration.

Change-Id: I5b16966434aa38f01e9fe092c1255c0f62bf8d83
Reviewed-on: https://review.streamsets.net/11096
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>