SDC-7751: Remove harcoded KafkaTestUtil.generateTestData from ProducerRunnable
In a galaxy far far away we built KafkaTestUtil that is tied to Kafka 0.8 APIs. It seems
that when adding support for Kafka 0.9, we've copy&pasted this class to SdcKafkaTestUtil
and make it abstract by removing all Kafka relevant calls down to subclasses that are
compiled against proper Kafka version.

However, we still use the KafkaTestUtil in ProducerRunnable and hence making a binary
dependency on certain Kafka 0.8 APIs! This patch transitions ProducerRunnable from using
directly KafkaTestUtil to using properly instantiated SdcKafkaTestUtil.

Change-Id: I0af4bef1943735fbd1b651d0c475b07857b8e9ea
Reviewed-on: https://review.streamsets.net/11103
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Jeff Evans <jeff@streamsets.com>