SDC-7262: Full Package Manager should not be available when running from parcel
Package manager have today two tasks:

* Managing (installing and removing) stage libraries
* Managing (installing and removing) external libraries

Since we ship the stage libraries as part of the parcel itself, the
first item doesn't make sense to be exposed on Cloudera Managed cluster.
The second item however does make sense.

This patch updates the UI code to detect if we're running under CM
management and if so, it will hide all stage-library relevant features
and displays only the external library management.

Change-Id: I1dd9b29423ab60538ce1f3269ad13682988908b6
Reviewed-on: https://review.streamsets.net/11117
Reviewed-by: Madhukar Devaraju <madhu@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>