SDC-7837: Rename Kafka configs to MapR Streams Configs
I would like to avoid sub-classing the main multi-topic consumer and
re-defining the configuration to match MapR naming scheme as we have
done that in the old origin and it turned out to be difficult to
maintain.

Rather then that, this patch changes the kafka specific wording on tab
and properties to be Kafka/MapR independent. We can update other
tooltips and such later as well.

Change-Id: I706e5deb3f1260ba06cac2980eaea8ba4a6a221e
Reviewed-on: https://review.streamsets.net/11493
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Jeff Evans <jeff@streamsets.com>
(cherry picked from commit 6a983dee7c7676bdd964d1285fb1482eafa1a38f)
Reviewed-on: https://review.streamsets.net/11496
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>