SDC-7946: Support bundle thread dump contains only up to 8 stack trace entries per thread
This is caused because the ThreadInfo.toString() have harcoded limit to
only print out up to 8 entries. That is sadly not enough in our code as
I for example needed more when troubleshooting failing tests.

This patch provides custom way of serializing thread dump info - it's
heavily inspired by the original ThreadInfo.toString() implementation,
so that tool written to understand it's format can read the custom
output as well. It does not however limit the stacktrace to only 8
entries.

Change-Id: I35ec5eff6b5c6d5a39aac9a7850396a08c03812d
Reviewed-on: https://review.streamsets.net/11680
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Ji Sun Kim <jisun@streamsets.com>