SDC-7974: Classpath: Add specific classpath whitelist files for stage libraries that needs it
This patch adds classspath  whitelist spec to each stage library that
needed it on my computer. Idea is to get the facility committed and
enabled by default, so that we can start fixing the issues as we go -
and more importantly make sure that new stage libraries will have a way
to detect classpath issues sooner rather then later.

Change-Id: I91a85c47bca52e7f0f2a45af5040f71ddc7ef07d
Reviewed-on: https://review.streamsets.net/11776
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Anthony McKeehan <anthony@streamsets.com>