SDC-7996: Service: Provide best effort implementation of Data Generator Processor
Idea of this patch is to provide implementation of a processor that
properly uses services, so that I can test that everything is working
properly when actually writing the internals of the data format service.

Tests of this component will be added later when the Data Format
Generator service will be done, so that it can be used for the actual
tests.

Change-Id: I04cdae30da43f46b343590fa3ff0838d5b521ab3
Reviewed-on: https://review.streamsets.net/11854
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Ji Sun Kim <jisun@streamsets.com>