SDC-5273: hive metadata doesn't produce rollup headers
This is fairly nasty bug that only shows up if both of the following
conditions are met:

1) There is more then one partition active for given table
2) A drift occurs

An example of how this could cause trouble is:

* Hive processor receives record with column /a for partition 1,
  HDFS destination opens file for partition 1.
* Hive processor receives record with column /a for partition 2,
  HDFS destination opens file for partition 2.
* Hive processor receives record with columns /a and /b for partition 1
  and hence sets the roll flag on this record. HDFS destination will
  close file for partition 1 and reopens a new one since the roll flag is
  present.
* Hive processor receives record with columns /a and /b for partition 2.
  Since this is not a schema drift any more, the roll flag will not be set
  on this record. Since HDFS destination have open file for partition 2
  with the old schema, the write operation will fail.

This patch enhances the internal partition cache by adding a new
property to hold set of partitions for which we need to roll the active
file. This set will contain all known partitions when a drift is
detected and each time we process at least one record for given
partition, we remove the partition from the set and at the same time set
the "roll" flag.

Change-Id: I196c258505fe09569bd958061386497d85358ecc
Reviewed-on: https://review.streamsets.net/11853
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>