SDC-8012: Support for auto uploading support bundles for pipeline crashes
This patch provides new configuration option for auto-bundle upload on
certain errors. Currently the only supported error is failing pipeline,
but we can add more later.

The functionality is disabled by default and user have to automatically
enable it if they are fine with us receiving bundle on every failed
pipeline.

Since the default pool for bundle generation is "1", enabling this
feature without additional tweaks on heavily used SDC is discouraged.

Change-Id: I6b481878b1248c426d3496f51a4ee813e401149d
Reviewed-on: https://review.streamsets.net/11948
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Junko Urata <junko@streamsets.com>