SDC-8084: JsonGeneratorOutputStream might not properly redact data
Rather then depending that the underlying JSON library will call the
write(byte []) at the interval we expect, this patch replaces the
implementation and depends on write(int). And caches all bytes until a
full line is available that is then redacted at one. This is the closest
possible behavior to other redaction that we're doing (line by line).

Change-Id: Icd00cfa5c9b1a05eb2abe6b4948c543ce54887bb
Reviewed-on: https://review.streamsets.net/12096
Reviewed-by: Bob Plotts <bob@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>