SDC-8129: Service: Move metrics methods to ProtoContext
This patch follows change from API-171 and moves all metric related
methods from Stage.Context to ProtoConfigurableEntity.Context so that
they can be easily reused between Services and Stages.

Currently the metrics from service shares the same namespace with Stage
metrics - this is undesirable, but since it's the same state as today
for data formats, it will be addressed in follow up work (it will have
impact on UI as well).

Change-Id: I7401ac6b5858db90da72402e7c17b069af8de75c
Reviewed-on: https://review.streamsets.net/12176
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>