SDC-8143: Serializer Avro Schema ID in all modes
This is actually my bad, when converting the code to support schema
registration from record header, I've added a check for schemaInHeader -
but that is actually not correct and we should serialize the schema id
in all cases when we have configured the schema registry. This matches
what we were doing before my changes.

Change-Id: I8924b7fa27b3c8fee406deda8bf13865bc0f103f
Reviewed-on: https://review.streamsets.net/12190
Reviewed-by: Adam Kunicki <adam@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>