SDC-8189: Propagate OnErrorRecordException to proper handler in Kudu Destnation
We already have code that properly detects if the table schema properly
allows NULLs for the column where we need to inset NULL and throws
OnErrorRecordException. Sadly the caller code is not catching this
exception and propagating it to the errorRecordHandler as it should.

Change-Id: I24de628c012252ee9e69c7d08a862fd933ad3ace
Reviewed-on: https://review.streamsets.net/12307
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Junko Urata <junko@streamsets.com>
(cherry picked from commit 05df784c691a0e7809dc24f9e972f15c780d9f29)
Reviewed-on: https://review.streamsets.net/12314
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>