SDC-8175: Lookup processors with enabled caching - add an option not to include default values in the cache
This patch introduces a new option on the shared cache config bean -
retry on cache miss. The idea is that negative lookups (value not found)
are not cached and will always be re-tried on next access.

Sadly the underlying Guava cache does not deal nicely with cache miss
events, so the patch is rather larger then it necessarily had to be and
introduces the Cache as always returning Optional to allow the "miss"
state.

All other lookups using the shared code are not using this new approach
yet as their code have to be transitioned to use the Optional class.
This will be done separately via separate patches.

Change-Id: I8ba6ea329ce1debe78c1052c9f4bdcc237773bf6
Reviewed-on: https://review.streamsets.net/12345
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Pat Patterson <pat@streamsets.com>
Reviewed-by: Junko Urata <junko@streamsets.com>