SDC-5544. Allow configuration of client cert/key for Salesforce connection
Add configuration for Keystore in all Salesforce stages. Uses an underlying
TlsConfigBean to leverage the existing Keystore code. Had to remove Apache
Fluent HTTP client as it does not allow the SSL socket factory to be
configured. Replaced with Jetty HTTP client as it is already used in the
Salesforce Streaming API client code.

Change-Id: Id2fd3ba7286a6c6dfbec6ffaa0c3c2e302a0db3c
Reviewed-on: https://review.streamsets.net/12464
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Jeff Evans <jeff@streamsets.com>