SDC-8095: JMS Consumer Throws SDC_RECORD_PARSER_00 for SDC Record
This one was a bit harder to uncover. I've added some various print outs
and noticed that the serialized message from data format library starts
with 0xA1 0x7B 0x22 0x68 and then the reader side sees 0xEF 0xBF 0xBD 0x7b
0x22 0x68 instead. E.g. somehow the 0xA1 was converted to 0xEFBFBD.
Doing a bit of a googling and it turns out that those three bytes are
special unicode placeholder for "invalid character".

Looking further in code I found out that we were considering SDC_JSON
format as a text format and explicitly casted it to String before
sending it to JMS. This is however wrong as SDC format is not a text one
- because we have that special byte at the begging that configures
whether the rest of the message is a special-json or kryo encoded data.

Change-Id: I36700d73b5bb7a7b68f2086b7a419ceaaf3eeb1b
Reviewed-on: https://review.streamsets.net/12545
Reviewed-by: Anthony McKeehan <anthony@streamsets.com>