SDC-8391: Explicitly exclude certain netty libraries in CDH stage libraries
This is a follow up on SDC-8354 that fixed Jackson and this patch fixes the
same problem with netty. To sum it up, there is a bug inside the assembly plugin
(MASSEMBLY-762) that makes the plugin not work properly with asterisk exclusions
and thus we need to exclude the jars manually.

Change-Id: Iddfc51d829902099821d5dbf2179bc3ec92abe4a
Reviewed-on: https://review.streamsets.net/12709
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>
(cherry picked from commit 7802448a94b552232710a2c9d4dd57f3a372b4be)
Reviewed-on: https://review.streamsets.net/12751
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>