SDC-8578: Add ability for stage library to express minimal SDC version
We're increasingly seeing random stage libraries showing up on the internet and some
of the SCH functionality is also delivered in forms of proprietary stage libraries. Today
the stage lib doesn't have any good way to declare a dependency on certain SDC version
which leads to all sort of random exceptions when one mistakenly loads given stage
library to wrong SDC version.

This patch introduces a new class capable of parsing version strings -
the logic is "borrowed" from SCH code base - but I've added a lot of
public APIs to slightly increase the readability of the code) and then
it plugs the class to the ClassLoaderStageLibraryTask so that each
stage library can expose a configuration on what is the minimal support
SDC version. If the SDC version is not new enough the start up procedure
will be terminated.

Change-Id: I39b59e111ea645614937db250df0f2f146e1e071
Reviewed-on: https://review.streamsets.net/13291
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Jeff Evans <jeff@streamsets.com>