SDC-8646. Remove expired TGTs and related service tickets when new TGT is received.
This patch removes old TGTs from the subject as soon as a new one is received on a
fresh login. This ensures that at no time do we have expired TGTs sitting around
in the subject's private credentials.

Previously, we removed TGTs only if they
had already expired when we re-login. Since we always re-login between 50 and 70%
of a TGT's validity, this would create a situation where a TGT would expire,
and we'd remove it only on the next login. Hadoop would try to create
service tickets with the first TGT it found, which could be the expired one - this
would fail until the next login when the expired TGT is removed. Example:
Ticket 1: Time of Issue: 0 Validity: 10 Credentials: Ticket 1
Ticket 2: Time of Issue: 7, Validity: 17 Credentials: Ticket 1, Ticket 2

Time 10: Ticket 1 is expired, but not removed. Any access to this ticket between
10 and 17 would cause pipeline failure. So we must remove it when Ticket 2 is
added to credentials.

This patch also removes service tickets associated with Ticket 1. This is because
the JDK expects that, once there is a new login, the first
ticket in the private credentials is the TGT. So we must remove all older service
tickets to get the new service tickets or the pipeline fails with an error saying
a TGT was expected.

Change-Id: I03df86a9a7f8315fd44d424ab12e026312d8a66b
Reviewed-on: https://review.streamsets.net/13434
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
(cherry picked from commit 8460a729e51fb4c2f8945c7978c8d3c565865c58)
Reviewed-on: https://review.streamsets.net/13607
Reviewed-by: Hari Shreedharan <hshreedharan@streamsets.com>