SDC-8636: Data Rules don't work in multithreaded pipelines because of NPE
The various structures are interconnected, so making them just ConcurrentX won't sadly solve the
underlying problem - the wrong index could be accessed at the wrong time,... I have a patch that
moves all of the interconnected structures into ThreadLocal which will move the computation to
each thread separately - and hence fixes the immediate issue.

As Virag Kothari mentioned though, this is a short-term fix - I feel that we need to have a
single place that will sample records across all batches (across all threads and all runners),
otherwise, the algorithm won't be working properly (or in an expected way). Doing the sampling only
per thread or per runner have the problem of skewing what data are observed by given thread or
runner. This won't be a huge trouble with let say HTTP Origin as data are passed both through
different threads or runners in a random way, but in other origins, like Kafka or JDBC Multi-table
origin, there are relationships between thread/runners that might lead to incorrect/unexpected
sampling outputs.

Change-Id: I9e88a6e951a72cb946395e61e20c292fa1273175
Reviewed-on: https://review.streamsets.net/13681
Reviewed-by: Virag Kothari <virag@streamsets.com>