SDC-4766. Spark 2.x Support for Cluster Mode
Version changes in pom files are based on the matrix in design doc.

So we'll have three cluster-bootstrap jars, kafka0.8, kafka0.10, and mapr.
This is necessary because classes and API are different betweem Kafka 0.8 and 0.10.
Instead it will be easy to drop Kafka 0.8 support when we want to near future.

This path includes fix for MapR. Adding security options to Spark job
is handled in SDC-6857.

Change-Id: Icc1912582e35381cd35826cd8a0b9b0828597bd7
Reviewed-on: https://review.streamsets.net/13825
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Hari Shreedharan <hshreedharan@streamsets.com>