SDC-8693: Service: On first pipeline auto-save any change to service config is not saved
This one was a bit more involved - to simplify joining data and metadata, I've
added a new $scope entity that is calculated when open the pipeline canvas -
$scope.detailPaneServices. One thing that I've missed originally is that when
the pipeline is saved for whatever reason, there is a different code path that
re-initializes the $scope variables. Since I have not updated originally this
different path, the pre-calculated entity were from the older pipeline version
and hence no change was detected.

Change-Id: I6865857e947f6c1a131c5384c55c356616e22a9b
Reviewed-on: https://review.streamsets.net/13737
Reviewed-by: Madhukar Devaraju <madhu@streamsets.com>