SDC-8721: Service: Introduce new maven module stagesupport
Commonlib is currently very fat library and I was hoping that stage libs
will be able to depend only on "common" after we introduce the
dataformat stage library. Sadly this is not entirely true as the
"common" itself have some undesirable dependencies such as jackson.

Hence I'm introducing a new maven module called 'stagesupport' - it's
what common was suppose to be, but without any extra dependencies. I'm
hoping that this is only a transitional state - once we will convert all
stages and stage libraries to use services and fully move dataformat
code to it's own stage library, I'm hoping that we will be able to
merge all common, commonlib and stagesupport into a single supporting
module that will be small in size and dependencies.

However for the time being in order to not touch stages that shouldn't
be touched as they are, we will need to have the trio of the modules.

Change-Id: Ic1f518df34c8bd29e3ef6d95afc76df2db5612e4
Reviewed-on: https://review.streamsets.net/13855
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>