SDC-8755: Time classpath validation execution
I got a feedback offline that the classpath validation can be "slow". In
my testing with all stagelibs it does take more then 2 seconds, but then
just loading stages from stagelibraries takes almost 30 seconds, so this
doesn't really seems like the performance hog. Nevertheless let's print
the time it takes to validate classpath, so that we can explore when
that number is too high.

Change-Id: I51b5e1946b34f6f45f6aa0966895436b0ca1b74e
Reviewed-on: https://review.streamsets.net/13913
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Loretta Chen <loretta@streamsets.com>