SDC-8937: StageRunner doesn't allow to run destroy() if initialization failed
I've noticed while debugging some troubles that our (test-based) StageRunner
will throw state exception if one tries to call runDestroy() without
successful finish of runInit(). This seems inconsistent with the main code
base, as we will actually call destroy() method on the component if it's init()
has failed.

This patch allows destroy() to be called even when the stage was freshly
created as that is consistent with what can happen on the normal (fully
fledge SDC) path.

Change-Id: I6bd14dc60c563ca699f254657cb702996ff128dc
Reviewed-on: https://review.streamsets.net/14429
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>