SDC-8953: Replace use of GlobPattern with GlobFilter in HdfsFileSystem
This is also coming from CDH 6 support. Hadoop 3 has changed the regular
expression library from Java to Google's RE2. Since there is no way how to
convert Google's Pattern class to Java's Pattern class, we either need a
lot of reflection or a separate implementation for each Hadoop version.
Rather then doing either the third option seems to be to utilize GlobFilter
itself that provides the same boolean answer that we need for the rest of
the code.

Change-Id: I11292b3de01ab6bbb1787ea414b8478ccad3fffd
Reviewed-on: https://review.streamsets.net/14454
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>