SDC-8978: Fully wire default interceptor creation to runtime
This patch wires all the runtime structures together so that the default
interceptors works end-to-end. Actually specifying what interceptors
should be run in non-default part is not covered by this patch. There is
quite a few upcoming improvements that I will need once this gets
merged, but I feel that this patch is large enough already and actually
works end  to end, so let's merge it as it stands.

Change-Id: Ic1eea511b2b22a82cfbfa8b9b83d001add72af9f
Reviewed-on: https://review.streamsets.net/14516
Reviewed-by: Jeff Evans <jeff@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>