SDC-8979: Make sure that Interceptors are called for Snapshots
It turns out that preview and snapshots have the same code path -
getting preview is about getting snapshot of a pipeline that runs in
"temporal" mode.

The way we make snapshot today is that we clone all incoming records
inside BatchMakerImpl. I did not want to change the BatchMakerImpl
yet, so I decided to run the interceptor twice - once for normal output
and second for the copied snapshot. Since both preview and snapshots
will run for one batch a time I feel that it's sufficient solution for
now.

Once we will get into metrics part of the project, we can reopen the
question if this is the right place for running the interceptors in the
first place and optimize the fact that they are running twice for
snapshots.

Change-Id: Ifc1e9cdd047daef790b4e0c9afa85a2cd2a8eef2
Reviewed-on: https://review.streamsets.net/14521
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Jeff Evans <jeff@streamsets.com>