SDC-8990. Add experimental parboiled-based parser to Oracle CDC origin
Adding a new PEG parser for the SQL format expected from LogMiner.
This is experimental, but seems to work in the testing done. Also
added tests to ensure that it can parse the SQL statements the
current one ANTLR-based one parses.

Based on some testing, the performance of this one is better.
Currently, it is experimental and marked as such. To use this parser,
the user must explicitly enable it in the "Advanced" tab in the
origin's configuration. This is to avoid regressions where the
origin is unable to parse SQL statements it was previously
able to parse.

Change-Id: Iaf8c2206278817eda004a1728c1ee7e76b11f9b0
Reviewed-on: https://review.streamsets.net/14527
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>
(cherry picked from commit 099743dc2c26dd5cd3fd7ff84daa1dfb7ed691e9)
Reviewed-on: https://review.streamsets.net/14719
Reviewed-by: Hari Shreedharan <hshreedharan@streamsets.com>
(cherry picked from commit 5fe80768b58396cb0c0d090488fa38846617e63a)
Reviewed-on: https://review.streamsets.net/14802