SDC-9101: Delegate: Add creator for the delegate
This is the last patch in the series to get the delegate infrastructure
working. It adds a new creator similarly as we have for services and
stages that is responsible for generating the new instance and a runtime
wrapper that will properly switch class loaders.

Eventually I would like to auto-generate the methods that needs to wrap
the classloader as the same action is required for services as well.
That is however a future exercise.

Change-Id: I13a4abdf5017b61b6a3fe3835c8df94ff8556d61
Reviewed-on: https://review.streamsets.net/14774
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>