SDC-9126. GCS origin reads only 1 batch from last file in queue.
This patch fixes the offset handling and selecting eligible files across sdc restarts on google cloud origin.
Previously we were not reading more than the first batch of the file because when we read the first batch of the the last file and next time produce method is called if the file queue is empty we were immediately throwing the no more data event and not reading data after that.

Change-Id: I8b56fb5b359884f4b934c5bb08aa2f1495469c1b
Reviewed-on: https://review.streamsets.net/14823
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>