SDC-9207: Service: S3 Destination still depends on data format configs
The reason for this dependency was to "hide" two configs when one uses
whole file format. Since stage is not allowed to depend on a service
config, we will have to display the options anyway and they will simply
go unused.

I have some thoughts how to refactore the whole file format a bit later,
but until then this will be the case.

Change-Id: Id573d5a4d1a97de3f4574ec943fefaf253bc1902
Reviewed-on: https://review.streamsets.net/14953
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>