SDC-9177. File watcher pipeline does not pick up new files
If batch wait time (sec) is set to a large number, we observe the follwing race condition.
When there are no files in the queue, readlock acquires for the batch wait time (sec) e.g. batch wait time (sec) may be set to 10 min, the readlock will be acquired for 10 min. Meanwhile, the writelock for adding the list of files to the queue is pending until the readlock is released.
We should release the readlock if we found the new files, pretty much immediately.

Change-Id: I615e937b86c00bf9b9d4a0c5ae9525177e3ca8c8
Reviewed-on: https://review.streamsets.net/14954
Reviewed-by: Hari Shreedharan <hshreedharan@streamsets.com>