SDC-9238: Run validation and upgrade during DetachedStage construction
Frankly this turned out to be much larger change that I've anticipated.
I had to move some shared code to a util class to make sure that it's
easily reusable in three different places. As part of this refactoring
I've also fixes few small bugs in the normal pipeline validator (too
small for an explicit JIRA).

Change-Id: I95c2fdf6a895623e5e3af4fb5c714d7dd9286ee2
Reviewed-on: https://review.streamsets.net/14998
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>