SDC-9242: Make Interceptor creation parametrizable
The original implementation for creating the interceptors only allowed
creating default interceptors and have not even provided a way to inject
all the important references (like blob store and configuration). It was
done as a proof of concept and hence the shortcut.

This implementation refactores that part of the code to expose a proper
interface to fully initialize the creator with all the right references.

Change-Id: Ifb1ef4c91c64bc78268404ebbbb26b48853a3ca3
Reviewed-on: https://review.streamsets.net/15006
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>