SDC-9243: Unify Runner.start methods
My apologies for a bit larger change - we currently have two different
start() methods calls that are used in various places. As I will need
to add a new argument to the start method for interceptors, it seen a
need to unify the different calls together so that it's easier to add a
new argument.

All this patch does is creating context for all information about start
and then propagating those changes everywhere. No other functional
aspect has been introduced in this patch.

Change-Id: I85a0119f1765311102036f0a5b7a622da0f3e5b6
Reviewed-on: https://review.streamsets.net/15007
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>