SDC-9245: Final wiring of interceptor configuration
Final wiring of everything together - now when Control Hub sends
interceptor configuration, the configs should be properly propagated all
the way to the PipelineBeanCreator.

There is a small catch that I will deal in with separate change - the
interceptor configuration is not currently serialized in the pipeline
status file. This means that if the pipeline auto-restarts, it will run
with the empty configuration (and thus default policies).

Change-Id: Iedcd0a546ad889f581f3f5f36ab38dcea82718eb
Reviewed-on: https://review.streamsets.net/15010
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>