SDC-9261: Do not validate input&output lanes for detached stage
This is an oversight on my side. When I tested the instantiation the
validation code wasn't there yet.

We already do a similar check for executors and destinations in case
that they are used as error, stats or pipeline lifecycle stage, so the
change should not be controversial - it's just skipping the same check
for processor as well.

Change-Id: I1e84a1cc671d9e41f1c2520ace70bddb044a5e80
Reviewed-on: https://review.streamsets.net/15056
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Jeff Evans <jeff@streamsets.com>